+++ src/index.py
@@ -1,5 +1,6 @@
 from tkinter import Tk
-from ui import UI
+from ui.ui import UI
+
 
 def main():
     window = Tk()
@@ -9,7 +10,6 @@ def main():
     ui_view.start()
 
     window.mainloop()
-
 
 
 if __name__ == "__main__":