+++ src/services/login_service.py
@@ -1,0 +1,6 @@
+from entities.user import User
+
+
+class Login:
+    def __init__(self) -> None:
+        pass