+++ dokumentaatio/changelog.md
@@ -5,4 +5,14 @@
 - lisätty .coveragerc ja changelog.md
 
 
+# Viikko 4
+
+- lisätty login_service
+- lisätty ui, repositories, entities ja services jako
+- lisätty user_repository, user.py 
+- lisätty build.py ?? sekä modattu database_connection.py
+
+
+
+